miienlp/tokens/requirements.txt
alive-progress==1.6.2
attrs==20.3.0
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.0.0/en_core_web_lg-3.0.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.0.0/en_core_web_trf-3.0.0-py3-none-any.whl
filelock==3.0.12
ftfy==5.9
idna==2.10
importlib-metadata==3.7.0
iniconfig==1.1.1
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
murmurhash==1.0.5
nltk==3.5
numpy==1.20.1
packaging==20.9
pandas==1.2.3
pathy==0.4.0
pluggy==0.13.1
preshed==3.0.5
py==1.10.0
pydantic==1.7.3
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
six==1.15.0
smart-open==3.0.0
spacy==3.0.3
spacy-alignments==0.7.2
spacy-legacy==3.0.1
spacy-transformers==1.0.1
srsly==2.4.0
thinc==8.0.1
tokenizers==0.9.4
toml==0.10.2
torch==1.8.0
torchcontrib==0.0.2
tqdm==4.58.0
transformers==4.2.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
wcwidth==0.2.5
zipp==3.4.0
